Prevent users from flooding us with reports
Fixes #6. This patch requires that any user GUID submit a report no
fewer than 30 minutes after their last report on the same escalator.

This includes a couple of minor cleanups, fixes the return of the full
escalator info (including new statuses) when an update succeeds, and
of course unit tests cover all of this.